6 files you always hoped but never totally believed would leave khan-exercises
Summary:
Port from [khan-exercises](https://github.com/Khan/khan-exercises/tree/master/utils)/.

Files were switched from requirejs syntax to node require syntax.

A couple utilities in kline.js were removed.
`kline.equal` was reverted to an older version for now
because it was an external contribution prior to the CLA.
`kline.intersect` was removed because it used a different
api format than the rest of utilities in `kline`

kmatrix was removed because it isn't used in Perseus and is a
lot of legacy code from earlier khan-exercises. We can add these
back when we need them if they match CLA requirements.

Test Plan:
none. I'll be adding tests and refactoring these to use
explicit underscore dependencies in a couple commits